requirements.txt
google-adk==1.18.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests>=2.32.4
